add pipeline metadata
